traceability_backend/requirements.txt
# ─── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ─── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ─── Validation & Config ───────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── QR Code ───────────────────────────────────────────────────────────────────
qrcode[pil]==7.4.2
Pillow==10.3.0

# ─── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ─── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ─── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
